+++ server/src/controllers/eventController.js
@@ -3,6 +3,7 @@ const { Event } = require("../db");
 //Creación de una cita:
 const createEvent = async (date, hour, purpose) => {
     const newEvent = await Event.create({ date, hour, purpose });
+    
     return newEvent;
 }
